Split user-defined types stuff into separate files for clarity
